+++ JWServer/SampleServer/SampleServerConfig.cpp
@@ -28,8 +28,8 @@ namespace jw
         return GetInt64("max-client-session-count");
     }
 
-    int32_t SampleServerConfig::GetTimerTickIntervalMilliSecond()
+    int64_t SampleServerConfig::GetTimerTickIntervalMilliSecond()
     {
-        return GetInt32("timer-tick-interval-milliSecond");
+        return GetInt64("timer-tick-interval-milliSecond");
     }
 }
